Refactor MapRoute and Request classes to utilize new Utils validation traits. Remove deprecated Validate trait and enhance type matching in route handling.
